backend/requirements.txt
jupyter
fastapi
pandas
dask[dataframe]
python-decouple
sqlalchemy
mysqlclient
uvicorn
jinja2
yfinance
pydantic
python-dotenv
